Fixes #35: Less strict filename filter when extracting natives
The suffix is "lib" for OSX (because that's what we have in the
Maven package name). It's also the suffix for the native library
files. Unfortunately the extensions are ".dylib" and ".jnilib".

Not requiring the dot when checking filenames is the less intrusive
option for fixing this. (The other option I can see being: extend the
lwjgl.os tuple with a new element: the list of extensions we want).